Associate with NRO files
This commit adds the ability to open NRO files directly from a file browser rather than the emulator itself.
